🪒 Razor: Flatten errors module hierarchy into single file
- Removed nested `src/errors/` directory structure.
- Merged `src/errors/assembly.rs` and `src/errors/mod.rs` directly into a single flat `src/errors.rs` file.
- Documented reductions in `.jules/razor.md`.
